Rely on BUILD_SHARED_LIBS instead of custom options
Instead of having your own option for choosing between static
and shared versions of the library, use cmake's built-in option
for this:
https://cmake.org/cmake/help/v3.0/variable/BUILD_SHARED_LIBS.html

Set -DBUILD_SHARED_LIBS=ON to get a shared library, and omit it
or set it to OFF to get a static one.
Can add one extra line to the cmake file to make the default be
shared. Makes most of the cmake crap go away.