xxh3: keys are xored, instead of added
In UMAC, keys are added, instead of xored.
According to its authors, that's because adds are 10% faster on their target platform
(a Pentium IV, this was early 2000).

In measurements with SSE/AVX, it makes no difference today.

However, additions make it possible to reason about delta between fields,
meaning it's easier to engineer a collision just knowing these deltas.
Xoring instead messes that up.
Since it's performance free, let's to that.