changed (again) formula for len:4-8
This one is no longer bijective (for len==8),
but produces expected nb of collisions,
suitable for 64-bit hash.

Previous variant was failing the DiffDist test,
when top bits where extracted from the hash :
the full hash value was bijective, hence no collision,
but grabing just a portion of it results in higher than normal collision rate.
This could be solved with another >> * serialized round,
but at this point, the algorithm's latency worsens (139M -> 115M).

Moving to mul128 is a lesser hit (139M->137M),
plus it fits nicely with the larger size variants,
re-using the same "core" mixer.