Add configuration to reroll most loops.
XXH_REROLL: Rerolls the following functions:
 - XXH32_finalize
 - XXH64_finalize
 - XXH3_64bitsWithSeed
 - XXH3_128bitsWithSeed
 - XXH3_accumulate (disables pragma)
Automatically enabled with __OPTIMIZE_SIZE__.

XXH_REROLL_XXH64: Specifically controls rerolling XXH64_finalize.
This is automatically defined on non-64-bit targets, as the unroll
makes the binary huge (it is the largest function in the program),
and it slows things down.

If XXH_REROLL is defined, this is ignored and the loop is rerolled
anyways.