Force -O2 on GCC + AVX2, document split load
GCC for AVX2 goes overboard on the unrolling with -O3, causing slower
code than MSVC and Clang.

We can override that with a pragma that forces GCC to use -O2 instead.

Note that GCC still generates the best scalar and SSE2 code with -O3.

I also mentioned the fact that GCC will split _mm256_loadu_si256 into
two instructions on a generic+avx2 target (which is an optimization that
only applies to the non-AVX2 Sandy and Ivy Bridge chips), and provide
the recommended flags.