[BUGFIX] Manually align XXH_malloc to 64 bytes.
Thanks to @RedSpah for finding this bug!

XXH3_state_t is manually 64-byte aligned on the stack, but not manually
aligned when dynamically allocated.

This caused an issue with where the state was assumed to be aligned
and it caused a segmentation fault.

Now, instead of somewhat pointlessly wrapping malloc and free, we write
a portable routine that overallocates and manually aligns the pointer
to 64 bytes.

There is no point in using posix_memalign or _mm_malloc when our
portability requirements would require this fallback anyways.