fix: don't interpret options in file list
When parsing arguments, the first non-option is treated as the start of
a list of file names.  However, arguments continue to be parsed, so any
later arguments that look like options will have a dual meaning as an
option and file name.  This is unexpected behavior, and here are two
solutions:

- Treat later option-like arguments as options
- Treat later option-like arguments as file names

Go with the latter solution for now because it is easier.  The former
solution is preferred because it is more flexible and similar to how
`shasum` parses its arguments.