Don't attempt to dispatch to AVX2 on GCC <= 4.9 (closes #473)
Although GCC 4.7 through 4.9 have AVX and AVX2 support and the necessary definitions, they aren't accessible to dispatcher code because you must turn on AVX/AVX2 mode to access them. We don't want to compile with that on as then the dispatcher code itself may get AVX/AVX2 optimizations and break.