Add QEMU/gcc matrix to ci.yml
This change adds support for older version of gcc cross compiler.
For further tweaking, I leave the matrix as just a bulky text table.
We'll be able to refactor it after we'll implement stable test.

For some reason, ubuntu-20.04 fails to retrieve `gcc-8-mips-linux-gnu`.
In this change we use ubuntu-18.04 for it.
We use ubuntu-20.04 for other gcc-8 tests though.