fix #627
attempt to fix a minor PVS Studio warning
which claims that passing a value by copy is worse for performance than a value by reference,
which, given that the value is 16 bytes long, is already silly just from the start,
but is even more so in this case as the function is short designed to be inlined by the compiler anyway.

Made the inlining objective more explicit, in an attempt to silence the static analyzer objection.

This is a blind fix, as I don't have PVS Studio to check.